requirements.txt
python-telegram-bot==8.1.1
-e git://github.com/NoMore201/googleplay-api.git#egg=gpapi